requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
librosa
matplotlib
